tests/wpt/web-platform-tests/dom/nodes/Element-childElementCount-dynamic-add.svg

<?xml version="1.0" encoding="UTF-8"?>
<svg xmlns="http://www.w3.org/2000/svg"
     xmlns:h="http://www.w3.org/1999/xhtml"
     version="1.1"
     width="100%" height="100%" viewBox="0 0 400 400">
<title>Dynamic Adding of Elements</title>
<h:script src="/resources/testharness.js"/>
<h:script src="/resources/testharnessreport.js"/>

<text x="200" y="40" font-size="25" fill="black" text-anchor="middle">Test of Dynamic Adding of Elements</text>
<text id="parentEl" x="200" y="70" font-size="20" fill="black" text-anchor="middle">The result of this test is
<tspan id="first_element_child" font-weight="bold">unknown.</tspan></text>

<h:script><![CDATA[
test(function() {
  var parentEl = document.getElementById("parentEl");
  var newChild = document.createElementNS("http://www.w3.org/2000/svg", "tspan");
  parentEl.appendChild(newChild);
  assert_equals(parentEl.childElementCount, 2)
})
]]></h:script>
</svg>
